Finished update implementation.
Initial commit of this resulted in problems in handling the net.IP.
This now contains a commit which changed the `IPAddresses` field of the
config struct into a slice of strings.

* Use pointers to certificateconfs

Very small but quite influential change. Without the pointers there is a
great risk of clashing configurations. Which usually is a cause for
great anguish and gnashing of teeth.

* Refactor to use own error type to distinguish recoverable situations

* Handle vault responses as strings

Also fix shadowed variable

* Handle non existing files in the beginning

* Refactor validity testing logic in update

* Add logging to vault

* Refactor inner variable naming for better readability

* Handle external command on update

* Add thread safe way of getting singleton vault client

* Extract read and update into methods for later use

* Add mutex to certificate retrieval

This should help prevent possible race conditions while fetching
certificates from vault.

* Prune away some dead code

* Handle errors to satisfy go-metalinter